feat: 삭제 확인 모달 시스템 모달로 변경

+++ app/admin/table/_components/mobile/post-table-toolbar.tsx
@@ -1,8 +1,6 @@
 import Check from "@/components/shared/check";
 import { Post } from "@/api/dto/post";
-import React, { useState } from "react";
 import { useDeletePosts } from "@/hooks/post/use-delete-posts";
-import TwoButtonModal from "@/app/_components/modal/two-button-modal";
 
 interface PostTableToolbarProps {
   posts: Post[];
@@ -16,7 +14,6 @@ const PostTableToolbar = ({
   setSelectedIds,
 }: PostTableToolbarProps) => {
   const { deletePosts } = useDeletePosts();
-  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
 
   // 전체 선택 여부 확인
   const isAllSelected =
@@ -33,9 +30,13 @@ const PostTableToolbar = ({
 
   // 선택 삭제 핸들러
   const handleDelete = async () => {
-    setIsDeleteModalOpen(false);
-    const success = await deletePosts(selectedIds);
-    console.log(success);
+    const confirmed = confirm(
+      "해당 게시글을 삭제하시나요?\n\n삭제된 게시글은 다시 복구할 수 없으며, 해당 카테고리 목록에서 제외됩니다.",
+    );
+    if (confirmed) {
+      const success = await deletePosts(selectedIds);
+      console.log(success);
+    }
   };
 
   return (
@@ -62,21 +63,11 @@ const PostTableToolbar = ({
         className={`pretendard-body-3 ${
           selectedIds.length === 0 ? "text-[#A7A9B4]" : "text-[#474953]"
         }`}
-        onClick={() => setIsDeleteModalOpen(true)}
+        onClick={handleDelete}
         disabled={selectedIds.length === 0}
       >
         선택 삭제
       </button>
-
-      {isDeleteModalOpen && (
-        <TwoButtonModal
-          title="해당 게시글을 삭제하시나요?"
-          desc="삭제된 게시글은 다시 복구할 수 없으며, 해당 카테고리 목록에서 제외됩니다."
-          buttonText="확인"
-          onClickFirstBtn={() => setIsDeleteModalOpen(false)}
-          onClickSecondBtn={handleDelete}
-        />
-      )}
     </div>
   );
 };
